`sct_deepseg`: Move seg function into `inference.py`
This better conforms to our API/CLI separation goals.
