+++ spinalcordtoolbox/deepseg/inference.py
@@ -1,0 +1,72 @@
+import logging
+from pathlib import Path
+
+from ivadomed import inference as imed_inference
+import nibabel as nib
+import numpy as np
+
+logger = logging.getLogger(__name__)
+
+
+def segment_and_average_volumes(model_paths, input_filenames, options):
+    """
+        Run `ivadomed.inference.segment_volume()` once per model, then average the outputs.
+
+        :param model_paths: A list of folder paths. The folders must contain:
+            (1) the model ('folder_model/folder_model.pt') to use
+            (2) its configuration file ('folder_model/folder_model.json') used for the training,
+            see https://github.com/neuropoly/ivadomed/wiki/configuration-file
+        :param input_filenames: list of image filenames (e.g. .nii.gz) to segment. Multichannel models require multiple
+            images to segment, i.e.., len(fname_images) > 1.
+        :param options: A dictionary containing optional configuration settings, as specified by the
+            ivadomed.inference.segment_volume function.
+
+        :return: list, list: List of nibabel objects containing the soft segmentation(s), one per prediction class, \
+            List of target suffix associated with each prediction
+    """
+    if not isinstance(model_paths, list):
+        raise TypeError("'model_paths' must be a list of model path strings.")
+    if not len(model_paths) > 0:
+        raise ValueError("'model_paths' must contain one or more model path strings.")
+
+    # Fetch the name of the model (to be used in logging)
+    name_model = Path(model_paths[0]).parts[-1]
+    logger.info(f"\nRunning inference for model '{name_model}'...")
+
+    # Perform inference once per model
+    nii_lsts, target_lsts = [], []
+    for path_model in model_paths:
+        if len(model_paths) > 1:  # We have an ensemble, so output messages to distinguish between seeds
+            name_seed = Path(path_model).parts[-2]
+            logger.info(f"\nUsing '{name_seed}'...")
+        nii_lst, target_lst = imed_inference.segment_volume(path_model, input_filenames, options=options)
+        nii_lsts.append(nii_lst)
+        target_lsts.append(target_lst)
+
+    # If we have a single model, skip averaging
+    if len(model_paths) == 1:
+        nii_lst = nii_lsts[0]
+    # Otherwise, we have a model ensemble, so average the image data
+    else:
+        logger.info(f"\nAveraging outputs across the ensemble for '{name_model}'...")
+        nii_lst = []
+        # NB: `nii_lsts` is a list of lists, with each sublist being the *per-model* predictions. Example:
+        #         [
+        #             [m1_prediction_1.nii.gz, m1_prediction_2.nii.gz, ...],  # model 1 predictions
+        #             [m2_prediction_1.nii.gz, m2_prediction_2.nii.gz, ...]   # model 2 predictions
+        #         ]
+        # So, we want to take: the average of "prediction_1", the average of "prediction_2", etc.
+        # To do this, we unpack + zip `nii_lists`, so that "prediction_N" files are grouped as "predictions".
+        for predictions in zip(*nii_lsts):
+            # Average the data for each output in the ensemble
+            data_stack = np.stack([pred.get_fdata() for pred in predictions], axis=0)
+            data_mean = np.mean(data_stack, axis=0)
+            # Take the first image's header to reuse for the averaged image
+            nii_header = predictions[0].header
+            # Create a new Nifti1Image containing the averaged output
+            nii_lst.append(nib.Nifti1Image(data_mean, header=nii_header, affine=nii_header.get_best_affine()))
+
+    # The 'targets' should be identical for each model, so just take the first
+    target_lst = target_lsts[0]
+
+    return nii_lst, target_lst
